Processed the refactoring of the CoMPAS core project.
