+++ app/src/main/java/org/lfenergy/compas/scl/data/rest/model/GetResponse.java
@@ -5,7 +5,7 @@
 package org.lfenergy.compas.scl.data.rest.model;
 
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
-import org.lfenergy.compas.scl.SCL;
+import org.lfenergy.compas.scl.model.SCL;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
