+++ repository-postgresql/src/test/java/org/lfenergy/compas/scl/data/repository/CompasSclDataPostgreSQLRepositoryTest.java
@@ -5,9 +5,9 @@ package org.lfenergy.compas.scl.data.repository;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.lfenergy.compas.scl.SCL;
 import org.lfenergy.compas.scl.data.model.SclType;
 import org.lfenergy.compas.scl.data.model.Version;
+import org.lfenergy.compas.scl.model.SCL;
 import org.mockito.InjectMocks;
 import org.mockito.junit.jupiter.MockitoExtension;
