+++ service/src/test/java/org/lfenergy/compas/scl/data/service/CompasSclDataServiceTest.java
@@ -6,15 +6,15 @@ package org.lfenergy.compas.scl.data.service;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.lfenergy.compas.commons.CompasExtensionsManager;
-import org.lfenergy.compas.scl.SCL;
-import org.lfenergy.compas.scl.THeader;
+import org.lfenergy.compas.scl.commons.CompasExtensionsManager;
 import org.lfenergy.compas.scl.data.model.ChangeSetType;
 import org.lfenergy.compas.scl.data.model.SclType;
 import org.lfenergy.compas.scl.data.model.Version;
 import org.lfenergy.compas.scl.data.repository.CompasSclDataRepository;
-import org.lfenergy.compas.scl.extensions.ObjectFactory;
-import org.lfenergy.compas.scl.extensions.TSclFileType;
+import org.lfenergy.compas.scl.extensions.model.ObjectFactory;
+import org.lfenergy.compas.scl.extensions.model.TSclFileType;
+import org.lfenergy.compas.scl.model.SCL;
+import org.lfenergy.compas.scl.model.THeader;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
@@ -22,8 +22,8 @@ import java.util.UUID;
 
 import static java.util.Collections.emptyList;
 import static org.junit.jupiter.api.Assertions.*;
-import static org.lfenergy.compas.scl.extensions.common.CompasExtensionsField.SCL_FILETYPE_EXTENSION;
-import static org.lfenergy.compas.scl.extensions.common.CompasExtensionsField.SCL_NAME_EXTENSION;
+import static org.lfenergy.compas.scl.extensions.commons.CompasExtensionsField.SCL_FILETYPE_EXTENSION;
+import static org.lfenergy.compas.scl.extensions.commons.CompasExtensionsField.SCL_NAME_EXTENSION;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
